refactor: 폴더 구조 & ãinvite클래스 4개의 클래스로 위임
