+++ backend/src/main/java/com/cgs/backend/websocket/service/invite/InviteResponseHandler.java
@@ -1,65 +1,27 @@
-
-
-package com.cgs.backend.websocket.service;
+package com.cgs.backend.websocket.service.invite;
 
 import com.cgs.backend.global.enums.UserStatus;
-import com.cgs.backend.websocket.dto.GameInviteRequest;
-import com.cgs.backend.websocket.dto.GameInviteResponse;
+import com.cgs.backend.websocket.dto.invite.GameInviteResponse;
+import com.cgs.backend.websocket.service.online.OnlineUserService;
 import com.cgs.backend.websocket.util.WebSocketEndpoint;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.redis.core.StringRedisTemplate;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
-import org.springframework.stereotype.Service;
+import org.springframework.stereotype.Component;
 
 import java.util.Set;
 
-@Service
-@Slf4j
+@Component
 @RequiredArgsConstructor
-public class GameInviteService {
+public class InviteResponseHandler {
 
     private final SimpMessagingTemplate messagingTemplate;
-    private final OnlineUserService onlineUserService;
     private final StringRedisTemplate redisTemplate;
+    private final OnlineUserService onlineUserService;
 
     private static final String PENDING_INVITE_KEY_PREFIX = "pending_invite:";
 
-    public void sendInvite(GameInviteRequest request) {
-        if (request.getFromUserId().equals(request.getToUserId())) {
-            messagingTemplate.convertAndSend(
-                    WebSocketEndpoint.userInviteResponse(request.getFromUserId()),
-                    new GameInviteResponse(false, "자기 자신을 초대할 수 없습니다.")
-            );
-            return;
-        }
-
-        UserStatus userStatus = onlineUserService.getUserStatus(request.getToUserId());
-
-        if (userStatus == UserStatus.OFFLINE) {
-            messagingTemplate.convertAndSend(
-                    WebSocketEndpoint.userInviteResponse(request.getFromUserId()),
-                    new GameInviteResponse(false, request.getToNickname() + "님은 오프라인입니다.")
-            );
-            return;
-        } else if (userStatus == UserStatus.IN_GAME) {
-            messagingTemplate.convertAndSend(
-                    WebSocketEndpoint.userInviteResponse(request.getFromUserId()),
-                    new GameInviteResponse(false, request.getToNickname() + "님은 게임 중입니다.")
-            );
-            return;
-        }
-
-        String key = PENDING_INVITE_KEY_PREFIX + request.getToUserId();
-        redisTemplate.opsForSet().add(key, request.getFromUserId());
-
-        messagingTemplate.convertAndSend(
-                WebSocketEndpoint.userInvite(request.getToUserId()),
-                request
-        );
-    }
-
-    public void processInviteResponse(GameInviteResponse response) {
+    public void handleInviteResponse(GameInviteResponse response) {
         String key = PENDING_INVITE_KEY_PREFIX + response.getToUserId();
 
         if (Boolean.TRUE.equals(response.getAccepted())) {
